requirements.txt
aiohttp
asyncpg
psycopg2-binary
python-dotenv
gitpython
rocketchat-API